app/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

aiohttp=3.7.4.post0=py38h497a2fe_0
anyio=3.3.0=pypi_0
argon2-cffi=20.1.0=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.9.3=pypi_0
bleach=4.0.0=pypi_0
bqplot=0.12.30=pypi_0
branca=0.4.2=pyhd8ed1ab_0
brotlipy=0.7.0=py38h497a2fe_1001
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.10.8=ha878542_0
cachetools=4.2.2=pyhd8ed1ab_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=8.0.1=py38h578d9bd_0
colour=0.1.5=pypi_0
cryptography=3.4.7=py38ha5dfef3_0
cycler=0.10.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.4.1=pypi_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.12=pypi_0
earthengine-api=0.1.275=pyhd8ed1ab_0
entrypoints=0.3=pypi_0
ffmpeg-python=0.2.0=pypi_0
flask=2.0.2=pyhd8ed1ab_0
folium=0.12.1=pypi_0
freetype=2.10.4=h0708190_1
future=0.18.2=py38h578d9bd_3
geeadd=0.5.5=pypi_0
geehydro=0.2.0=pyh9f0ad1d_0
geemap=0.8.18=pypi_0
geetools=0.6.11=pypi_0
geocoder=1.38.1=pypi_0
geojson=2.5.0=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
google-api-core=1.31.1=pyhd8ed1ab_0
google-api-python-client=1.12.8=pyhd3deb0d_0
google-auth=1.34.0=pyh6c4a22f_0
google-auth-httplib2=0.1.0=pyhd8ed1ab_0
google-cloud-core=1.7.2=pyh6c4a22f_0
google-cloud-storage=1.19.0=py_0
google-crc32c=1.1.2=py38h8838a9a_0
google-resumable-media=1.3.3=pyh6c4a22f_0
googleapis-common-protos=1.53.0=py38h578d9bd_0
googledrivedownloader=0.4=pypi_0
grpcio=1.38.1=py38hdd6454d_0
gunicorn=20.1.0=py38h578d9bd_1
httplib2=0.19.1=pyhd8ed1ab_0
httplib2shim=0.0.3=pypi_0
idna=3.1=pyhd3deb0d_0
ipyevents=2.0.1=pypi_0
ipyfilechooser=0.4.4=pypi_0
ipykernel=6.2.0=pypi_0
ipyleaflet=0.14.0=pypi_0
ipynb-py-convert=0.4.6=pypi_0
ipython=7.26.0=py38he5a9106_0
ipython_genutils=0.2.0=py_1
ipytree=0.2.1=pypi_0
ipywidgets=7.6.3=pypi_0
itsdangerous=2.0.1=pyhd8ed1ab_0
jbig=2.1=h7f98852_2003
jedi=0.18.0=py38h578d9bd_2
jinja2=3.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-core=4.7.1=pypi_0
jupyter-server=1.10.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=pypi_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=10_openblas
libcblas=3.9.0=10_openblas
libcrc32c=1.1.1=h9c3ff4c_2
libdeflate=1.7=h7f98852_5
libffi=3.3=h58526e2_2
libgcc-ng=11.1.0=hc902ee8_6
libgfortran-ng=11.1.0=h69a702a_6
libgfortran5=11.1.0=h6c583b3_6
libgomp=11.1.0=hc902ee8_6
liblapack=3.9.0=10_openblas
libopenblas=0.3.17=pthreads_h8fe5266_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.17.2=h780b84a_1
libstdcxx-ng=11.1.0=h56837e0_6
libtiff=4.3.0=hf544144_1
libwebp-base=1.2.0=h7f98852_2
logzero=1.7.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=pypi_0
mss=6.1.0=pypi_0
multidict=5.1.0=py38h497a2fe_1
nbclient=0.5.4=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pypi_0
notebook=6.4.3=pypi_0
numpy=1.21.1=py38h9894fe3_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1l=h7f98852_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.1=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.3.1=py38h8e6f84c_0
pip=21.2.2=pyhd8ed1ab_0
plotly=5.2.1=pypi_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.19=pyha770c72_0
protobuf=3.17.2=py38h709712a_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pycrs=1.0.2=pypi_0
pygithub=1.55=pypi_0
pygments=2.9.0=pyhd8ed1ab_0
pyjwt=2.1.0=pypi_0
pynacl=1.4.0=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.18.0=pypi_0
pyshp=2.1.3=pypi_0
pysocks=1.7.1=py38h578d9bd_3
python=3.8.10=h49503c6_1_cpython
python-box=5.4.0=pypi_0
python-dateutil=2.8.2=pypi_0
python-dotenv=0.19.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=22.2.1=pypi_0
ratelim=0.1.6=pypi_0
readline=8.1=h46c0cb4_0
requests=2.26.0=pyhd8ed1ab_0
requests-unixsocket=0.2.0=pypi_0
rsa=4.7.2=pyh44b312d_0
sankee=0.0.5=pypi_0
send2trash=1.8.0=pypi_0
setuptools=49.6.0=py38h578d9bd_3
simplejson=3.17.3=py38h497a2fe_0
six=1.16.0=pyh6c4a22f_0
smmap=4.0.0=pypi_0
sniffio=1.2.0=pypi_0
soupsieve=2.2.1=pypi_0
sqlite=3.36.0=h9cd32fc_0
tenacity=8.0.1=pypi_0
terminado=0.11.1=pypi_0
testpath=0.5.0=pypi_0
tk=8.6.10=h21135ba_1
tornado=6.1=pypi_0
traitlets=5.0.5=py_0
traittypes=0.2.1=pypi_0
typing-extensions=3.10.0.0=hd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
uritemplate=3.0.1=py_0
urllib3=1.26.6=pyhd8ed1ab_0
voila=0.2.10=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
websocket-client=1.2.1=pypi_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
whitebox=1.5.2=pypi_0
whiteboxgui=0.3.0=pypi_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h516909a_1
yarl=1.6.3=py38h497a2fe_2
zlib=1.2.11=h516909a_1010
zstd=1.5.0=ha95c52a_0
